feat: rewrote the docker module with a specific result type with some of the std::Result implementation to manipulate docker responses
