+++ src/lib.rs
@@ -1,6 +1,3 @@
-#[macro_use]
-mod macros;
-
 pub mod container;
 pub mod docker;
 pub mod error;
@@ -17,12 +14,4 @@ pub(crate) fn serialize_as_json<T: serde::Serialize, S: serde::Serializer>(
     s.serialize_str(
         &serde_json::to_string(t).map_err(|e| serde::ser::Error::custom(e.to_string()))?,
     )
-}
-
-#[macro_export]
-macro_rules! debug_print {
-    ($($arg:tt)*) => {
-        #[cfg(debug_assertions)]
-        println!($($arg)*);
-    };
 }
