--- src/macros.rs
@@ -1,5 +1,0 @@
-macro_rules! docker_response {
-    ($resp:expr) => {
-        crate::docker::DockerResponse::from($resp)
-    };
-}
